customizable MotionDetector
